+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Free memory whenever `LightXML` is used
 ### Other
 - Fill some small gaps
+- Ability to do some routing calculations with updated weights based on already walked and whether it's counting towards the goal
 
 ## v0.1.2 (31st of December 2024)
 ### Bugfixes